+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### [1.2] - 2020-11-11 
+#### Added
+- Improve way of calling rx scripts
+- PROGDIR as path for HippoPlayerIR.config
+
 ### [1.1] - 2019-09-23 
 #### Added
 - A proper way of serial port setup (now it doesn't relay on default settings from Preferences)